refactor(#620): create settings module for IO
store logic for the settings schema in the new module
