+++ src/config_ninja/__init__.py
@@ -1,28 +1,43 @@
 """.. include:: ../../README.md
 
 # Navigation
-- `config_ninja.contrib` for supported backends:
-    - `config_ninja.contrib.appconfig`
-    - `config_ninja.contrib.local`
-- `config_ninja.cli` for commands and CLI documentation
-- `config_ninja.systemd` for `systemd` integration
+
+## `config_ninja.contrib`
+
+For supported backends.
+
+### `config_ninja.contrib.appconfig`
+
+Integrate with the AWS AppConfig service.
+
+### `config_ninja.contrib.local`
+
+Use a local file as the backend.
+
+## `config_ninja.cli`
+
+Commands and CLI documentation.
+
+## `config_ninja.settings`
+
+For settings and configuration.
+
+## `config_ninja.systemd`
+
+Integration with `systemd`.
 """  # noqa: D415
 
 from __future__ import annotations
 
 __version__ = '0.0.0'
 
-from pathlib import Path
-
-from pyspry import Settings
+from config_ninja.settings import DEFAULT_PATHS
+from config_ninja.settings import load as load_settings
+from config_ninja.settings import resolve_path as resolve_settings_path
 
 __all__ = ['DEFAULT_SETTINGS_PATHS', 'load_settings', 'resolve_settings_path']
 
-DEFAULT_SETTINGS_PATHS = [
-    Path.cwd() / 'config-ninja-settings.yaml',
-    Path.home() / 'config-ninja-settings.yaml',
-    Path('/etc/config-ninja/settings.yaml'),
-]
+DEFAULT_SETTINGS_PATHS = DEFAULT_PATHS
 """Check each of these locations for `config-ninja`_'s settings file.
 
 The following locations are checked (ordered by priority):
@@ -31,21 +46,5 @@ The following locations are checked (ordered by priority):
 2. `~/config-ninja-settings.yaml`
 3. `/etc/config-ninja/settings.yaml`
 
-.. _config-ninja: https://bryant-finney.github.io/config-ninja/config_ninja.html
+.. _config-ninja: https://config-ninja.readthedocs.io/home.html
 """
-
-SETTINGS_PREFIX = 'CONFIG_NINJA'
-
-
-def load_settings(path: Path) -> Settings:
-    """Load the settings from the given path."""
-    return Settings.load(path, SETTINGS_PREFIX)
-
-
-def resolve_settings_path() -> Path:
-    """Return the first path in `DEFAULT_SETTINGS_PATHS` that exists."""
-    for path in DEFAULT_SETTINGS_PATHS:
-        if path.is_file():
-            return path
-
-    raise FileNotFoundError('Could not find config-ninja settings', DEFAULT_SETTINGS_PATHS)
